Run gateone as the web user
Run gateone as the web user, both for security and practical reasons.

Practically, this fixes a problem where in multi-node stacks, the SSH
connect script would be launched as root, and if it downloaded the
student stack SSH key before the XBlock, the XBlock would then proceed
to error out with a permissions exception.